requirements.txt
python-telegram-bot==20.3
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
pytz==2023.3
firebase-admin==6.1.0
python-dotenv==1.0.0